Implement validator to disallow using unsafe functions from Kernel
Note we deleted the validator that was checking for defmodule because it
is now covered by this new validator.
