+++ lib/live_view_demo/sandbox/command_validator.ex
@@ -9,9 +9,9 @@ defmodule LiveViewDemo.Sandbox.CommandValidator do
   @type ast :: Macro.t()
   @callback validate(ast()) :: :ok | {:error, String.t()}
 
-  alias LiveViewDemo.Sandbox.{AllowedElixirModules, DefmodulesAbsence, ErlangModulesAbsence}
+  alias LiveViewDemo.Sandbox.{AllowedElixirModules, ErlangModulesAbsence, SafeKernelFunctions}
 
-  @ast_validator_modules [DefmodulesAbsence, AllowedElixirModules, ErlangModulesAbsence]
+  @ast_validator_modules [SafeKernelFunctions, AllowedElixirModules, ErlangModulesAbsence]
 
   def safe_command?(command) do
     ast = Code.string_to_quoted!(command)
