+++ test/live_view_demo/sandbox/command_validator_test.exs
@@ -18,14 +18,14 @@ defmodule LiveViewDemo.Sandbox.CommandValidatorTest do
       assert {:error, "Invalid modules: [:File]"} == CommandValidator.safe_command?(command)
     end
 
+    test "returns :error when using an invalid Kernel function" do
+      assert {:error, "It is allowed to invoke only safe Kernel functions: [:apply]"} ==
+               CommandValidator.safe_command?("apply(Enum, :count, [[]])")
+    end
+
     test "returns :error when using an Erlang module" do
       assert {:error, "It is not allowed to invoke non-Elixir modules: [:lists]"} ==
                CommandValidator.safe_command?(":lists.last([5])")
-    end
-
-    test "returns :error when using defmodule" do
-      assert {:error, "Defining new modules is not allowed. Do not use `defmodule`."} ==
-               CommandValidator.safe_command?("defmodule Test, do: nil")
     end
   end
 end
